+++ src/app/components/EditModal.tsx
@@ -1,4 +1,3 @@
-'use client';
 import React from 'react';
 import { EditModalProps } from '../types';
